--- src/test/java/GetOrderTest.java
@@ -1,74 +1,0 @@
-import configs.BaseTest;
-import io.qameta.allure.Step;
-import io.restassured.http.ContentType;
-import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import requests.LoginRequest;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.*;
-
-public class GetOrderTest extends BaseTest {
-
-    private static String accessToken;
-
-    @BeforeAll
-    public static void setup() {
-        String email = "hilokea@yandex.ru";
-        String password = "kassian";
-        accessToken = loginAndGetToken(email, password);
-    }
-
-    @Step("Логин пользователя и получение accessToken")
-    private static String loginAndGetToken(String email, String password) {
-        LoginRequest loginRequest = new LoginRequest(email, password);
-        Response loginResponse = given()
-                .contentType(ContentType.JSON)
-                .body(loginRequest)
-                .when()
-                .post("/api/auth/login");
-
-        loginResponse.then().statusCode(200);
-        return loginResponse.jsonPath().getString("accessToken");
-    }
-
-    @Test
-    @DisplayName("Получение заказов авторизованным пользователем")
-    public void getOrdersWithAuthShouldReturn200() {
-        getOrdersWithAuth(accessToken);
-    }
-
-    @Step("Отправка запроса на получение заказов с токеном авторизации")
-    private void getOrdersWithAuth(String token) {
-        given()
-                .header("Authorization", token)
-                .when()
-                .get("/api/orders")
-                .then()
-                .statusCode(200)
-                .body("success", equalTo(true))
-                .body("orders", notNullValue())
-                .body("orders.size()", greaterThanOrEqualTo(0))
-                .body("total", greaterThanOrEqualTo(0))
-                .body("totalToday", greaterThanOrEqualTo(0));
-    }
-
-    @Test
-    @DisplayName("Получение заказов неавторизованным пользователем")
-    public void getOrdersWithoutAuthShouldReturn401() {
-        getOrdersWithoutAuth();
-    }
-
-    @Step("Отправка запроса на получение заказов без авторизации")
-    private void getOrdersWithoutAuth() {
-        given()
-                .when()
-                .get("/api/orders")
-                .then()
-                .statusCode(401)
-                .body("success", equalTo(false))
-                .body("message", equalTo("You should be authorised"));
-    }
-}
