+++ src/test/java/tests/EditUserTest.java
@@ -1,11 +1,12 @@
-import com.fasterxml.jackson.databind.ser.Serializers;
+package tests;
+
 import configs.BaseTest;
 import io.qameta.allure.Step;
 import io.restassured.http.ContentType;
-import io.restassured.response.Response;
 import org.junit.jupiter.api.*;
-import requests.LoginRequest;
 import requests.SignUpRequest;
+import steps.AuthSteps;
+import steps.CreateUserSteps;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
@@ -13,6 +14,8 @@ import static org.hamcrest.Matchers.*;
 public class EditUserTest extends BaseTest {
 
     private static String accessToken;
+    private static CreateUserSteps createUserSteps;
+    private static AuthSteps authSteps;
 
     @BeforeAll
     public static void setUp() {
@@ -20,33 +23,8 @@ public class EditUserTest extends BaseTest {
         String password = "adb123";
         String name = "leCroissant";
 
-        registerUser(email, password, name);
-        accessToken = loginAndGetToken(email, password);
-    }
-
-    @Step("Регистрация пользователя с email: {0}, name: {2}")
-    public static void registerUser(String email, String password, String name) {
-        SignUpRequest signUpRequest = new SignUpRequest(email, password, name);
-        given()
-                .contentType(ContentType.JSON)
-                .body(signUpRequest)
-                .when()
-                .post("/api/auth/register")
-                .then()
-                .statusCode(anyOf(is(200), is(403))); // 403 если уже зарегистрирован
-    }
-
-    @Step("Логин пользователя с email: {0}")
-    public static String loginAndGetToken(String email, String password) {
-        LoginRequest loginRequest = new LoginRequest(email,password);
-        Response loginResponse = given()
-                .contentType(ContentType.JSON)
-                .body(loginRequest)
-                .when()
-                .post("/api/auth/login");
-
-        loginResponse.then().statusCode(200);
-        return loginResponse.jsonPath().getString("accessToken");
+        createUserSteps.registerUser(email, password, name);
+        accessToken = authSteps.loginAndGetAccessToken(email, password);
     }
 
     @Test
@@ -89,6 +67,11 @@ public class EditUserTest extends BaseTest {
         updateUserWithoutAuth("{\"password\": \"noauthpass\"}");
     }
 
+    @AfterAll
+    public static void tearDown() {
+        createUserSteps.deleteUser(accessToken);
+    }
+
     @Step("Попытка обновления пользователя без авторизации с телом запроса: {0}")
     public void updateUserWithoutAuth(String requestBody) {
         given()
@@ -100,22 +83,5 @@ public class EditUserTest extends BaseTest {
                 .statusCode(401)
                 .body("success", is(false))
                 .body("message", equalTo("You should be authorised"));
-    }
-
-    @AfterAll
-    public static void tearDown() {
-        deleteUser(accessToken);
-    }
-
-    @Step("Удаление пользователя")
-    public static void deleteUser(String token) {
-        if (token != null) {
-            given()
-                    .header("Authorization", token)
-                    .when()
-                    .delete("/api/auth/user")
-                    .then()
-                    .statusCode(202);
-        }
     }
 }
